+++ src/Factory/Nyholm.php
@@ -9,7 +9,6 @@ use Nyholm\Psr7\Factory\Psr17Factory;
 use Nyholm\Psr7Server\ServerRequestCreator;
 use Override;
 use Psr\Http\Message\ServerRequestInterface;
-use Throwable;
 
 /** @api */
 class Nyholm extends AbstractFactory
@@ -18,22 +17,21 @@ class Nyholm extends AbstractFactory
 
 	public function __construct()
 	{
-		try {
-			$factory = new Psr17Factory();
-			$this->factory = $factory;
-			$this->responseFactory = $factory;
-			$this->streamFactory = $factory;
-			$this->requestFactory = $factory;
-			$this->serverRequestFactory = $factory;
-			$this->uploadedFileFactory = $factory;
-			$this->uriFactory = $factory;
-
-			// @codeCoverageIgnoreStart
-		} catch (Throwable) {
-			throw new RuntimeException('Install laminas/laminas-diactoros');
-
-			// @codeCoverageIgnoreEnd
+		if (!class_exists(Psr17Factory::class) || !class_exists(ServerRequestCreator::class)) {
+			throw new RuntimeException(
+				'Install nyholm/psr7 and nyholm/psr7-server to use the default '
+				. 'PSR-7 factory, or pass a custom Factory implementation to the App constructor',
+			);
 		}
+
+		$factory = new Psr17Factory();
+		$this->factory = $factory;
+		$this->responseFactory = $factory;
+		$this->streamFactory = $factory;
+		$this->requestFactory = $factory;
+		$this->serverRequestFactory = $factory;
+		$this->uploadedFileFactory = $factory;
+		$this->uriFactory = $factory;
 	}
 
 	#[Override]